test(better-auth): keep basic customer creation coverage
